refactor: Rename lotto purchaser class - Change name from lottoBuyer to lottoPurchaser
