+++ src/main/java/lotto/model/lottoPurchaser/InMemoryLottoRepository.java
@@ -1,4 +1,4 @@
-package lotto.model.lottoBuyer;
+package lotto.model.lottoPurchaser;
 
 import lotto.model.lotto.Lotto;
